refactor(inlayhints): вынести индекс вызовов в отдельный класс и устранить замечания Sonar
