+++ src/test/java/com/github/_1c_syntax/bsl/languageserver/inlayhints/PlatformMethodCallInlayHintSupplierUnitTest.java
@@ -87,7 +87,7 @@ class PlatformMethodCallInlayHintSupplierUnitTest {
     // when
     var hints = supplier.getInlayHints(documentContext, params);
 
-    // then — L96 return List.of().
+    // then — при отсутствии AST подсказок нет.
     assertThat(hints).isEmpty();
   }
